Update signup form
