+++ src/Screens/Signup/styles.tsx
@@ -1,8 +1,24 @@
 import {StyleSheet} from 'react-native';
-
 const styles = StyleSheet.create({
-  screen: {
+  titleText: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    color: '#333',
+  },
+  paragraph: {
+    marginVertical: 8,
+    lineHeight: 20,
+  },
+  container: {
     flex: 1,
+    padding: 20,
+  },
+  input: {
+    borderWidth: 1,
+    borderColor: '#ddd',
+    padding: 10,
+    fontSize: 18,
+    borderRadius: 6,
   },
 });
